ref(debuginfo): Move dwarf logic from Object to own trait
